// src/types/game.types.ts
export enum GameStatus {
  MENU = 'menu',
  PLAYING = 'playing',
  PAUSED = 'paused',
  GAME_OVER = 'game_over',
  LEVEL_TRANSITION = 'level_transition'
}

export enum DifficultyLevel {
  EASY = 'easy',
  NORMAL = 'normal',
  HARD = 'hard'
}

export enum GameMode {
  CLASSIC = 'classic',
  TIME_ATTACK = 'time_attack',
  ENDLESS = 'endless',
  DAILY_CHALLENGE = 'daily_challenge'
}

export enum Grade {
  BRONZE = 'bronze',
  SILVER = 'silver',
  GOLD = 'gold',
  DIAMOND = 'diamond',
  MASTER = 'master'
}